Added composite test to hopefully quiet pipeline error
